Mar. 26. 2023 - FE 재작업 필요
회원가입 영역 변수값 즉시 반영 안됨
